Update googleurl to the latest version from Chromium
The version used is cc5912fc759fc64d7b8553cd9bb912f2a26c0cc7 from
Mon Oct 3 10:33:09 2022 +0000

+++ url/url_canon_stdstring.h
@@ -1,4 +1,4 @@
-// Copyright 2013 The Chromium Authors. All rights reserved.
+// Copyright 2013 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -46,7 +46,7 @@ class COMPONENT_EXPORT(URL) StdStringCanonOutput : public CanonOutput {
   // Must be called after writing has completed but before the string is used.
   void Complete();
 
-  void Resize(int sz) override;
+  void Resize(size_t sz) override;
 
  protected:
   // `str_` is not a raw_ptr<...> for performance reasons (based on analysis of
